[refactor] onCanvasSizeChanged 함수 추가
