+++ app/src/main/java/com/squirtles/musicroad/detail/components/music/visualizer/soundeffect/DrawSoundBar.kt
@@ -18,11 +18,11 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundBarConstants.STROKE_WIDTH
-import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConstants.GRADIENT_RADIUS_RATIO
-import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConstants.OFFSET_ANGLE
-import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConstants.animatedGradientRadius
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.GRADIENT_RADIUS_RATIO
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.OFFSET_ANGLE
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.animatedGradientRadius
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.onCanvasSizeChanged
 import com.squirtles.musicroad.ui.theme.White
-import kotlin.math.min
 
 /* Bar 형태 원형 시각화 */
 @Composable
@@ -49,10 +49,14 @@ internal fun DrawSoundBar(
     val angleStep = 360f / audioData.size
 
     LaunchedEffect(canvasSize) {
-        adjustedRadius =
-            if (radius.value == 0f) (min(canvasSize.width, canvasSize.height) / 2f) * radiusRatio
-            else radius.value
-        maxBarHeight = (canvasSize.height / 4).toFloat()
+        onCanvasSizeChanged(
+            width = canvasSize.width,
+            height = canvasSize.height,
+            radius = radius.value,
+            radiusRatio = radiusRatio,
+            onRadiusCalculated = { adjustedRadius = it },
+            onMaxBarHeightCalculated = { maxBarHeight = it }
+        )
     }
 
     Canvas(
