+++ app/src/main/java/com/squirtles/musicroad/detail/components/music/visualizer/soundeffect/DrawSoundWaveFill.kt
@@ -22,10 +22,10 @@ import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
-import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConstants.OFFSET_ANGLE
-import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConstants.animatedGradientRadius
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.OFFSET_ANGLE
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.animatedGradientRadius
+import com.squirtles.musicroad.detail.components.music.visualizer.soundeffect.DrawSoundEffectConfigs.onCanvasSizeChanged
 import com.squirtles.musicroad.ui.theme.White
-import kotlin.math.min
 
 /* Wave 형태 원형 시각화 */
 @Composable
@@ -49,10 +49,14 @@ internal fun DrawSoundWaveFill(
     val animatedGradientRadius = animatedGradientRadius(FastOutSlowInEasing)
 
     LaunchedEffect(canvasSize) {
-        adjustedRadius =
-            if (radius.value == 0f) (min(canvasSize.width, canvasSize.height) / 2f) * radiusRatio
-            else radius.value
-        maxBarHeight = (canvasSize.height / 4).toFloat()
+        onCanvasSizeChanged(
+            width = canvasSize.width,
+            height = canvasSize.height,
+            radius = radius.value,
+            radiusRatio = radiusRatio,
+            onRadiusCalculated = { adjustedRadius = it },
+            onMaxBarHeightCalculated = { maxBarHeight = it }
+        )
         center = Offset(x = canvasSize.width / 2f, y = canvasSize.height / 2f)
         holePath = holePath.apply {
             addOval(
